Re-structured the queueserver module to be more BITS-like.
