+++ docs/how_to_guides/environment_configuration.rst
@@ -1,0 +1,28 @@
+Environment & Configuration
+===========================
+
+This how-to guide covers configuring a python environment to run
+Haven, Firefly, and the Haven Queueserver.
+
+Environmental Variables
+-----------------------
+
+To properly configure an environment for running Haven, Firefly, and
+the Queueserver, configure the following environmental variables.
+
+``HAVEN_CONFIG_DIR``
+  Path to the folder holding local configuration. This will also be used
+  for files automatically generated by e.g. the
+  Queueserver. E.g. ``~/bluesky``.
+``QSERVER_ZMQ_CONTROL_ADDRESS``
+  The ZeroMQ host and port to be used for clients to connect to the
+  queueserver. E.g. ``tcp://hostname.anl.gov:60615``
+``QSERVER_ZMQ_CONTROL_ADDRESS_FOR_SERVER``
+  The ZeroMQ host and port to be used by the queueserver
+  itself. E.g. ``tcp://*:60615``
+``QSERVER_ZMQ_INFO_ADDRESS``
+  The ZeroMQ host and port to be used for clients to connect to the
+  queueserver. E.g. ``tcp://hostname.anl.gov:60615``
+``QSERVER_ZMQ_INFO_ADDRESS_FOR_SERVER``
+  The ZeroMQ host and port to be used by the queueserver
+  itself. E.g. ``tcp://*:60625``
