+++ src/haven/iconfig_testing.toml
@@ -23,10 +23,6 @@ beamline = "255-ID-C"
 # to determine how to interact with the queue.
 
 [queueserver]
-control_host = "localhost"
-control_port = "60615"
-info_host = "localhost"
-info_port = "60625"
 redis_addr = "localhost:6379"
 
 [kafka]
