+++ src/haven_qserver/local_config.yml
@@ -1,0 +1,13 @@
+# Bluesky-Queueserver start-re-manager configuration
+# use:
+#   queue-monitor &
+#   ./qserver/qs_host.sh start
+#
+#   or
+#   cd ./qserver
+#   start-re-manager --config=./qs-config.yml
+#
+# https://blueskyproject.io/bluesky-queueserver/manager_config.html
+
+network:
+    zmq_control_addr: tcp://*:60631
