+++ src/haven_qserver/qs_config.yml
@@ -1,0 +1,47 @@
+# Bluesky-Queueserver start-re-manager configuration
+# use:
+#   queue-monitor &
+#   ./qserver/qs_host.sh start
+#
+#   or
+#   cd ./qserver
+#   start-re-manager --config=./qs-config.yml
+#
+# https://blueskyproject.io/bluesky-queueserver/manager_config.html
+
+network:
+    zmq_publish_console: true
+    # Per-beamline config: handled by CLI options
+    # redis_addr: localhost:6379
+    # redis_name_prefix: qs_default
+    
+    # Per-beamline config: handled by environmental variables
+    # zmq_control_addr: tcp://*:60615
+    # zmq_info_addr: tcp://*:60645
+    # zmq_info_addr: tcp://*:60625
+    
+
+operation:
+    # choices: SILENT QUIET NORMAL VERBOSE
+    console_logging_level: NORMAL
+
+    # emergency_lock_key: custom_lock_key
+
+    print_console_output: true
+
+    # choices: NEVER ENVIRONMENT_OPEN ALWAYS
+    update_existing_plans_and_devices: ENVIRONMENT_OPEN
+
+    # choices: NEVER ON_REQUEST ON_STARTUP
+    user_group_permissions_reload: ON_STARTUP
+
+startup:
+    startup_module: haven_qserver.startup
+    # Per-beamline config: handled by CLI options
+    # existing_plans_and_devices_path: ${BLUESKY_DIR}/queueserver_existing_plans_and_devices.yaml
+    # user_group_permissions_path: ${BLUESKY_DIR}/queueserver_user_group_permissions_bro.yaml
+
+worker:
+    use_ipython_kernel: true
+    # ipython_kernel_ip: auto
+    ipython_matplotlib: qt5
